requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.1
python-socketio==5.8.0
python-engineio==4.7.1
eventlet==0.33.3